Event route renamed as eventlogs
