+++ NEWS
@@ -1,3 +1,7 @@
+24.02.2021 Version 0.0.30
+Event route renamed as eventlogs
+
+
 23.02.2021 Version 0.0.29
 Excel Import feature added into dashboard.
 Default permission setting algorithm upgraded. New collections will inherit project's permissions. Similarly, fields will inherit collection's permissions. For data routes, if parentCollectionID is not defined, then inherit collection's permissions.
