+++ app.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 const express = require('express');
 const morgan = require('morgan');
 const rateLimit = require('express-rate-limit');
-const helmet = require('helmet');
+// const helmet = require('helmet');
 const mongoSanitize = require('express-mongo-sanitize');
 const xss = require('xss-clean');
 const hpp = require('hpp');
@@ -17,7 +17,7 @@ const globalErrorHandler = require('./controllers/errorController');
 const collectionsRouter = require('./routes/collectionsRoutes');
 const fieldsRouter = require('./routes/fieldsRoutes');
 const userRouter = require('./routes/userRoutes');
-const eventRouter = require('./routes/eventRoutes');
+const eventLogRouter = require('./routes/eventLogRoutes');
 const serverRouter = require('./routes/serverRoutes');
 const groupRouter = require('./routes/groupRoutes');
 const userGroupRouter = require('./routes/userGroupRoutes');
@@ -119,7 +119,8 @@ app.use('/api/v1/projects', projectsRouter);
 app.use('/api/v1/collections', collectionsRouter);
 app.use('/api/v1/fields', fieldsRouter);
 app.use('/api/v1/users', userRouter);
-app.use('/api/v1/events', eventRouter);
+// app.use('/api/v1/events', eventRouter);
+app.use('/api/v1/eventlogs', eventLogRouter);
 app.use('/api/v1/groups', groupRouter);
 app.use('/api/v1/server', serverRouter);
 app.use('/api/v1/usergroups', userGroupRouter);
