+++ routes/eventLogRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const authController = require('../controllers/authController');
-const eventController = require('../controllers/eventController');
+const eventLogController = require('../controllers/eventLogController');
 
 const router = express.Router();
 
@@ -8,7 +8,7 @@ router.use(authController.isLoggedIn);
 router.use(authController.requireLogin);
 router.use(authController.setDefPerms);
 
-router.route('/').get(eventController.getAllEvents);
-router.route('/:id').get(eventController.getEvent);
+router.route('/').get(eventLogController.getAllEventLogs);
+router.route('/:id').get(eventLogController.getEventLog);
 
 module.exports = router;
